bugfix in TessellationTable temporary buffer actually too small
    vk validation fixes

Script sync from SHA: 5cf4cc7c7b238c7d98d656dc01d1e3ecd0cdf8ee

+++ src/tessellation_table.cpp
@@ -1,5 +1,5 @@
 /*
-* Copyright (c) 2024-2025, NVIDIA CORPORATION.  All rights reserved.
+* Copyright (c) 2024-2026, NVIDIA CORPORATION.  All rights reserved.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
-* SPDX-FileCopyrightText: Copyright (c) 2024-2025, NVIDIA CORPORATION.
+* SPDX-FileCopyrightText: Copyright (c) 2024-2026, NVIDIA CORPORATION.
 * SPDX-License-Identifier: Apache-2.0
 */
 
@@ -182,7 +182,7 @@ void TessellationTable::initTemplates(Resources& res, uint32_t positionTruncateB
                                VMA_ALLOCATION_CREATE_MAPPED_BIT | VMA_ALLOCATION_CREATE_HOST_ACCESS_SEQUENTIAL_WRITE_BIT);
 
   nvvk::Buffer templateAddressesBuffer;
-  res.m_allocator.createBuffer(templateAddressesBuffer, sizeof(uint32_t) * tessellation_table::max_configs * 2,
+  res.m_allocator.createBuffer(templateAddressesBuffer, sizeof(uint64_t) * tessellation_table::max_configs * 2,
                                VK_BUFFER_USAGE_ACCELERATION_STRUCTURE_BUILD_INPUT_READ_ONLY_BIT_KHR, VMA_MEMORY_USAGE_CPU_ONLY,
                                VMA_ALLOCATION_CREATE_MAPPED_BIT | VMA_ALLOCATION_CREATE_HOST_ACCESS_SEQUENTIAL_WRITE_BIT);
